Change int to uint64 in domain.go

+++ domain/domain.go
@@ -12,7 +12,7 @@ type Product struct {
 }
 
 type ProductService interface {
-	GetProducts(start int, num int, sku string, barcode string, fields []string) ([]Product, error)
+	GetProducts(start uint64, num uint64, sku string, barcode string, fields []string) ([]Product, error)
 	GetProduct(id int, fields []string) (Product, error)
 	AddProduct(product Product) (int, error)
 	UpdateProduct(id int, product Product) error
@@ -20,7 +20,7 @@ type ProductService interface {
 }
 
 type ProductRepository interface {
-	GetProducts(start int, num int, sku string, barcode string) ([]Product, error)
+	GetProducts(start uint64, num uint64, sku string, barcode string) ([]Product, error)
 	GetProduct(id int) (Product, error)
 	AddProduct(product Product) (int, error)
 	UpdateProduct(id int, product Product) error
